add more code to handle downloading item status
