+++ pkg/triggers/starrqueue/lidarr.go
@@ -2,14 +2,12 @@ package starrqueue
 
 import (
 	"context"
-	"strings"
 	"time"
 
 	"github.com/Notifiarr/notifiarr/pkg/apps"
 	"github.com/Notifiarr/notifiarr/pkg/triggers/common"
 	"github.com/Notifiarr/notifiarr/pkg/triggers/data"
 	"github.com/Notifiarr/notifiarr/pkg/website"
-	"golift.io/starr/lidarr"
 )
 
 const TrigLidarrQueue common.TriggerName = "Storing Lidarr instance %d queue."
@@ -76,41 +74,4 @@ func (c *cmd) setupLidarr() bool {
 	}
 
 	return enabled
-}
-
-func (c *cmd) getFinishedItemsLidarr(_ context.Context) itemList { //nolint:cyclop
-	stuck := make(itemList)
-
-	for idx, app := range c.Apps.Lidarr {
-		ci := website.GetClientInfo()
-		if !app.Enabled() || ci == nil || !ci.Actions.Apps.Lidarr.Stuck(idx+1) {
-			continue
-		}
-
-		item := data.GetWithID("lidarr", idx)
-		if item == nil || item.Data == nil {
-			continue
-		}
-
-		queue, _ := item.Data.(*lidarr.Queue)
-		instance := idx + 1
-		stuckapp := stuck[instance]
-
-		for _, item := range queue.Records {
-			if s := strings.ToLower(item.Status); s != completed && s != warning &&
-				s != failed && s != errorstr && item.ErrorMessage == "" && len(item.StatusMessages) == 0 {
-				continue
-			}
-
-			stuckapp.Queue = append(stuckapp.Queue, item)
-		}
-
-		stuckapp.Name = c.Apps.Lidarr[idx].Name // this should be safe.
-		stuck[instance] = stuckapp
-
-		c.Debugf("Checking Lidarr (%d) Queue for Stuck Items, queue size: %d, stuck: %d",
-			instance, len(queue.Records), len(stuck[instance].Queue))
-	}
-
-	return stuck
 }
