+++ pkg/triggers/starrqueue/sonarr.go
@@ -2,14 +2,12 @@ package starrqueue
 
 import (
 	"context"
-	"strings"
 	"time"
 
 	"github.com/Notifiarr/notifiarr/pkg/apps"
 	"github.com/Notifiarr/notifiarr/pkg/triggers/common"
 	"github.com/Notifiarr/notifiarr/pkg/triggers/data"
 	"github.com/Notifiarr/notifiarr/pkg/website"
-	"golift.io/starr/sonarr"
 )
 
 const TrigSonarrQueue common.TriggerName = "Storing Sonarr instance %d queue."
@@ -78,46 +76,4 @@ func (c *cmd) setupSonarr() bool {
 	}
 
 	return enable
-}
-
-func (c *cmd) getFinishedItemsSonarr(_ context.Context) itemList { //nolint:cyclop
-	stuck := make(itemList)
-
-	for idx, app := range c.Apps.Sonarr {
-		ci := website.GetClientInfo()
-		if !app.Enabled() || ci == nil || !ci.Actions.Apps.Sonarr.Stuck(idx+1) {
-			continue
-		}
-
-		item := data.GetWithID("sonarr", idx)
-		if item == nil || item.Data == nil {
-			continue
-		}
-
-		queue, _ := item.Data.(*sonarr.Queue)
-		instance := idx + 1
-		stuckapp := stuck[instance]
-		// repeatStomper is used to collapse duplicate download IDs.
-		repeatStomper := make(map[string]*sonarr.QueueRecord)
-
-		for _, item := range queue.Records {
-			if s := strings.ToLower(item.Status); s != completed && s != warning &&
-				s != failed && s != errorstr && item.ErrorMessage == "" && len(item.StatusMessages) == 0 {
-				continue
-			} else if repeatStomper[item.DownloadID] != nil {
-				continue
-			}
-
-			repeatStomper[item.DownloadID] = item
-			stuckapp.Queue = append(stuckapp.Queue, item)
-		}
-
-		stuckapp.Name = c.Apps.Sonarr[idx].Name // this should be safe.
-		stuck[instance] = stuckapp
-
-		c.Debugf("Checking Sonarr (%d) Queue for Stuck Items, queue size: %d, stuck: %d",
-			instance, len(queue.Records), len(stuck[instance].Queue))
-	}
-
-	return stuck
 }
