+++ pkg/triggers/starrqueue/stuckitems.go
@@ -3,62 +3,15 @@ package starrqueue
 import (
 	"context"
 	"fmt"
-	"time"
+	"strings"
 
-	"github.com/Notifiarr/notifiarr/pkg/triggers/common"
+	"github.com/Notifiarr/notifiarr/pkg/triggers/data"
 	"github.com/Notifiarr/notifiarr/pkg/website"
+	"golift.io/starr/lidarr"
+	"golift.io/starr/radarr"
+	"golift.io/starr/readarr"
+	"golift.io/starr/sonarr"
 )
-
-/* This file contains the procedures to send stuck download queue items to notifiarr. */
-
-// Action contains the exported methods for this package.
-type Action struct {
-	cmd *cmd
-}
-
-type cmd struct {
-	*common.Config
-}
-
-const (
-	// How often to check starr apps for queue list when stuck items is enabled.
-	stuckDuration = 5 * time.Minute
-	// How often to check starr apps for queue list when finished items is enabled.
-	finishedDuration = time.Minute
-	// This is the max number of queued items to inspect/send.
-	queueItemsMax = 100
-)
-
-const (
-	errorstr  = "error"
-	failed    = "failed"
-	warning   = "warning"
-	completed = "completed"
-)
-
-const TrigStuckItems common.TriggerName = "Sending cached stuck items to website."
-
-// New configures the library.
-func New(config *common.Config) *Action {
-	return &Action{cmd: &cmd{Config: config}}
-}
-
-// Run initializes the library.
-func (a *Action) Create() {
-	lidarr := a.cmd.setupLidarr()
-	radarr := a.cmd.setupRadarr()
-	readarr := a.cmd.setupReadarr()
-	sonarr := a.cmd.setupSonarr()
-
-	if lidarr || radarr || readarr || sonarr {
-		a.cmd.Add(&common.Action{
-			Name: TrigStuckItems,
-			Fn:   a.cmd.sendStuckQueues,
-			C:    make(chan website.EventType, 1),
-			T:    time.NewTicker(stuckDuration),
-		})
-	}
-}
 
 // StuckItems sends the stuck queues items for all apps.
 // Does not fetch fresh data first, uses cache.
@@ -66,30 +19,7 @@ func (a *Action) StuckItems(event website.EventType) {
 	a.cmd.Exec(event, TrigStuckItems)
 }
 
-// listItem is data formatted for sending a json payload to the website.
-type listItem struct {
-	Name  string        `json:"name"`
-	Queue []interface{} `json:"queue"`
-}
-
-// itemList stores an instance->queue map.
-type itemList map[int]listItem
-
-func (i itemList) Len() int {
-	count := 0
-
-	for _, v := range i {
-		count += len(v.Queue)
-	}
-
-	return count
-}
-
-func (i itemList) Empty() bool {
-	return i.Len() < 1
-}
-
-// sendStuckQueues gathers the stuck quues from cache and sends them.
+// sendStuckQueues gathers the stuck queue from cache and sends them.
 func (c *cmd) sendStuckQueues(ctx context.Context, event website.EventType) {
 	lidarr := c.getFinishedItemsLidarr(ctx)
 	radarr := c.getFinishedItemsRadarr(ctx)
@@ -121,4 +51,157 @@ func (c *cmd) sendStuckQueues(ctx context.Context, event website.EventType) {
 			Sonarr:  sonarr,
 		},
 	})
+}
+
+func (c *cmd) getFinishedItemsLidarr(_ context.Context) itemList { //nolint:cyclop
+	stuck := make(itemList)
+
+	for idx, app := range c.Apps.Lidarr {
+		ci := website.GetClientInfo()
+		if !app.Enabled() || ci == nil || !ci.Actions.Apps.Lidarr.Stuck(idx+1) {
+			continue
+		}
+
+		item := data.GetWithID("lidarr", idx)
+		if item == nil || item.Data == nil {
+			continue
+		}
+
+		queue, _ := item.Data.(*lidarr.Queue)
+		instance := idx + 1
+		stuckapp := stuck[instance]
+
+		for _, item := range queue.Records {
+			if s := strings.ToLower(item.Status); s != completed && s != warning &&
+				s != failed && s != errorstr && item.ErrorMessage == "" && len(item.StatusMessages) == 0 {
+				continue
+			}
+
+			stuckapp.Queue = append(stuckapp.Queue, item)
+		}
+
+		stuckapp.Name = c.Apps.Lidarr[idx].Name // this should be safe.
+		stuck[instance] = stuckapp
+
+		c.Debugf("Checking Lidarr (%d) Queue for Stuck Items, queue size: %d, stuck: %d",
+			instance, len(queue.Records), len(stuck[instance].Queue))
+	}
+
+	return stuck
+}
+
+func (c *cmd) getFinishedItemsRadarr(_ context.Context) itemList { //nolint:cyclop
+	stuck := make(itemList)
+
+	for idx, app := range c.Apps.Radarr {
+		ci := website.GetClientInfo()
+		if !app.Enabled() || ci == nil || !ci.Actions.Apps.Radarr.Stuck(idx+1) {
+			continue
+		}
+
+		item := data.GetWithID("radarr", idx)
+		if item == nil || item.Data == nil {
+			continue
+		}
+
+		queue, _ := item.Data.(*radarr.Queue)
+		instance := idx + 1
+		stuckapp := stuck[instance]
+
+		for _, item := range queue.Records {
+			if s := strings.ToLower(item.Status); s != completed && s != warning &&
+				s != failed && s != errorstr && item.ErrorMessage == "" && len(item.StatusMessages) == 0 {
+				continue
+			}
+
+			stuckapp.Queue = append(stuckapp.Queue, item)
+		}
+
+		stuckapp.Name = c.Apps.Radarr[idx].Name // this should be safe.
+		stuck[instance] = stuckapp
+
+		c.Debugf("Checking Radarr (%d) Queue for Stuck Items, queue size: %d, stuck: %d",
+			instance, len(queue.Records), len(stuck[instance].Queue))
+	}
+
+	return stuck
+}
+
+func (c *cmd) getFinishedItemsReadarr(_ context.Context) itemList { //nolint:cyclop
+	stuck := make(itemList)
+
+	for idx, app := range c.Apps.Readarr {
+		ci := website.GetClientInfo()
+		if !app.Enabled() || ci == nil || !ci.Actions.Apps.Readarr.Stuck(idx+1) {
+			continue
+		}
+
+		item := data.GetWithID("readarr", idx)
+		if item == nil || item.Data == nil {
+			continue
+		}
+
+		queue, _ := item.Data.(*readarr.Queue)
+		instance := idx + 1
+		stuckapp := stuck[instance]
+
+		for _, item := range queue.Records {
+			if s := strings.ToLower(item.Status); s != completed && s != warning &&
+				s != failed && s != errorstr && item.ErrorMessage == "" && len(item.StatusMessages) == 0 {
+				continue
+			}
+
+			stuckapp.Queue = append(stuckapp.Queue, item)
+		}
+
+		stuckapp.Name = c.Apps.Readarr[idx].Name // this should be safe.
+		stuck[instance] = stuckapp
+
+		c.Debugf("Checking Readarr (%d) Queue for Stuck Items, queue size: %d, stuck: %d",
+			instance, len(queue.Records), len(stuck[instance].Queue))
+	}
+
+	return stuck
+}
+
+func (c *cmd) getFinishedItemsSonarr(_ context.Context) itemList { //nolint:cyclop
+	stuck := make(itemList)
+
+	for idx, app := range c.Apps.Sonarr {
+		ci := website.GetClientInfo()
+		if !app.Enabled() || ci == nil || !ci.Actions.Apps.Sonarr.Stuck(idx+1) {
+			continue
+		}
+
+		cacheItem := data.GetWithID("sonarr", idx)
+		if cacheItem == nil || cacheItem.Data == nil {
+			continue
+		}
+
+		queue, _ := cacheItem.Data.(*sonarr.Queue)
+		instance := idx + 1
+		stuckapp := stuck[instance]
+		// repeatStomper is used to collapse duplicate download IDs.
+		repeatStomper := make(map[string]*sonarr.QueueRecord)
+
+		for _, item := range queue.Records {
+			if s := strings.ToLower(item.Status); s != completed && s != warning &&
+				s != failed && s != errorstr && item.ErrorMessage == "" && len(item.StatusMessages) == 0 {
+				continue
+			} else if repeatStomper[item.DownloadID] != nil {
+				continue
+			}
+
+			repeatStomper[item.DownloadID] = item
+			stuckapp.Queue = append(stuckapp.Queue, item)
+		}
+
+		stuckapp.Name = c.Apps.Sonarr[idx].Name // this should be safe.
+		stuck[instance] = stuckapp
+
+		c.Debugf("Checking Sonarr (%d) Queue for Stuck Items, queue size: %d, stuck: %d",
+			instance, len(queue.Records), len(stuck[instance].Queue))
+	}
+
+	return stuck
 }
